+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - table can filter in or out keys (considering the first list represents the headers)
 - progress-bar prints something on a dumb terminal
 - colors and styles are enabled everywhere, including on "dumb" (non interactive) terminals like Emacs' Slime.
+- <2024-11-25> adapt the columns' widths to te terminal-width
 
 ## CL ANSI Term 0.1.2